requirements.txt
streamlit
faiss-cpu
python-dotenv
sentence-transformers
PyPDF2
numpy
google-generativeai
